smtp and refactoring of sendEmail into reportGen
