+++ server/reportGenerator/package.json
@@ -26,6 +26,8 @@
 	},
 	"dependencies": {
 		"aws-lambda": "^1.0.6",
-		"aws-sdk": "^2.809.0"
+		"aws-sdk": "^2.809.0",
+		"dotenv": "^10.0.0",
+		"nodemailer": "^6.6.3"
 	}
 }
