--- server/sendEmail/.eslintignore
@@ -1,5 +1,0 @@
-# don't ever lint node_modules
-node_modules
-
-# don't lint our output build
-dist
