--- server/sendEmail/jest.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  roots: ['<rootDir>/src'],
-  testMatch: ['**/?(*.)test.ts'],
-  transform: {
-    '^.+\\.ts$': 'ts-jest'
-  }
-}
