--- server/sendEmail/package.json
@@ -1,34 +1,0 @@
-{
-	"name": "vtaa",
-	"version": "1.0.0",
-	"scripts": {
-		"fast-compile": "tsc",
-		"compile": "eslint . --ext .ts && tsc",
-		"lint": "eslint . --ext .ts",
-		"test": "jest",
-		"format": "prettier --write \"**/*.{js,jsx,json,ts,md,yml}\""
-	},
-	"devDependencies": {
-		"@types/aws-lambda": "^8.10.81",
-		"@types/jest": "^26.0.24",
-		"@types/node": "^14.14.13",
-		"@types/nodemailer": "^6.4.4",
-		"@typescript-eslint/eslint-plugin": "^4.9.1",
-		"@typescript-eslint/parser": "^4.9.1",
-		"eslint": "^7.15.0",
-		"eslint-config-prettier": "^6.10.1",
-		"eslint-plugin-prettier": "^3.1.0",
-		"jest": "^27.0.6",
-		"moq.ts": "^6.4.0",
-		"prettier": "^2.0.1",
-		"ts-jest": "^27.0.4",
-		"ts-node": "^9.1.1",
-		"typescript": "^4.1.3"
-	},
-	"dependencies": {
-		"aws-lambda": "^1.0.6",
-		"aws-sdk": "^2.978.0",
-		"dotenv": "^10.0.0",
-		"nodemailer": "^6.6.3"
-	}
-}
