--- server/sendEmail/tsconfig.json
@@ -1,15 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "es2017",
-		"module": "commonjs",
-		"outDir": "./dist",
-		"strict": true,
-		"typeRoots": ["node_modules/@types"],
-		"types": ["node", "jest"],
-		"inlineSourceMap": true,
-		"esModuleInterop": true,
-		"skipLibCheck": true,
-		"forceConsistentCasingInFileNames": true
-	},
-	"exclude": ["node_modules/"]
-}
